+++ src/clients/data/database/clients.service.ts
@@ -1,0 +1,18 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class ClientsService {
+  
+  findAll() {
+    return `This action returns all clients`;
+  }
+
+  findOne(id: number) {
+    return `This action returns a #${id} client`;
+  }
+
+
+  remove(id: number) {
+    return `This action removes a #${id} client`;
+  }
+}
